fix[appc]: Download/uncompress to temporary directory
Instead of downloading or uncompressing directly to
`/var/cache/univention-appcenter/appcenter.software-univention.de/$ver/.all.tar`
use `.tmp.tar` as a temporary file. Verify and extract from there before
moving it to the final location at the end.

The time stamp of `.all.tar` is used determine, if `.apps.$lang.json`
should be updated. That process depends one the successful extraction of
the TAR as otherwise the directory is incomplete. This results in Apps
not being known, which may lead to installed Apps being displayed as
"unknown" or even removed.

Bug #54452