Bug #54467: move search widget into UDM syntax classes
* get rid of udm/values: it is either not required as we can put the
  same information into "udm/properties" and the request delivers the
  same as udm/syntax/choices.
* adapt MixedInput so that is capable of every widget not only TextBox, ComboBox and CheckBox.

Note: we add "search_defaults" as dict (not separate or tuple) to the
properties because the UMC store cannot handle arrays in top level.

TODO: don't use an empty array as indicator for the widget. Instead
assign it explicitly and allow a default value for the ComboBox values.

TODO: is {search_defaults: {default_value: ..., widget: ...}} a good choice?